refactor: move Ballots to neuron-detail
