+++ frontend/svelte/src/lib/components/neuron-detail/Ballots/BallotSummary.svelte
