+++ frontend/svelte/src/lib/components/neuron-detail/Ballots/Ballots.svelte
